requirements.txt
flask==2.2.0
flask[async]
werkzeug==2.2.0
numpy==2.0.0
pandas==1.4.0
plotly==5.5.0
requests==2.28.2
websocket-client==1.5.1
Flask-Session==0.5.0
redis==4.5.1
gunicorn==20.1.0
uvicorn==0.22.0
uvicorn[standard]
Flask-Caching==2.0.2
python-dotenv==1.0.0 